Working on build-w32 and build-w64 scripts

+++ scripts/build-w32.sh
@@ -190,7 +190,7 @@ END_OF_README
  LDFLAGS="-L${LIBUSB_1_0_BIN_PATH}/MinGW32/dll"
 
 
-make install
+make -j 4 install
 
 mkdir -p ${ZIP_DIR}/bin ${ZIP_DIR}/lib/msvc ${ZIP_DIR}/lib/gcc ${ZIP_DIR}/include ${ZIP_DIR}/doc
 cp -a src/libhamlib.def ${ZIP_DIR}/lib/msvc/libhamlib-4.def; todos ${ZIP_DIR}/lib/msvc/libhamlib-4.def
@@ -222,12 +222,23 @@ cp -a /usr/i686-w64-mingw32/lib/libwinpthread-1.dll ${ZIP_DIR}/bin/.
 cp -a ${LIBUSB_1_0_BIN_PATH}/MinGW32/dll/libusb-1.0.dll ${ZIP_DIR}/bin/libusb-1.0.dll
 
 # Required for MinGW with GCC 6.3
+FILE= /usr/lib/gcc/i686-w64-mingw32/6.3-posix/libgcc_s_sjlj-1.dll
+if test -f "$FILE"; then
 cp -a /usr/lib/gcc/i686-w64-mingw32/6.3-posix/libgcc_s_sjlj-1.dll ${ZIP_DIR}/bin/libgcc_s_sjlj-1.dll
+cp -a ${FILE} ${ZIP_DIR}/bin/libgcc_s_sjlj-1.dll
+fi
+
+# Required for MinGW with GCC 8.3
+FILE= /usr/lib/gcc/i686-w64-mingw32/8.3-posix/libgcc_s_sjlj-1.dll
+if test -f "$FILE"; then
+cp -a ${FILE} ${ZIP_DIR}/bin/libgcc_s_sjlj-1.dll
+fi
 
 ## Need VC++ free toolkit installed (default Wine directory installation shown)
 # Path for 2003 version of Visual C++ Toolkit
-#-( cd ${ZIP_DIR}/lib/msvc/ && wine ~/.wine/drive_c/Program\ Files/Microsoft\ Visual\ C++\ Toolkit\ 2003/bin/link.exe /lib /machine:i386 /def:libhamlib-4.def )
-# Path for Current version of Visual Studio
+#( cd ${ZIP_DIR}/lib/msvc/ && wine ~/.wine/drive_c/Program\ Files/Microsoft\ Visual\ C++\ Toolkit\ 2003/bin/link.exe /lib /machine:i386 /def:libhamlib-4.def )
+# Path for 2017 version of Visual Studio
 ( cd ${ZIP_DIR}/lib/msvc/ && wine ~/.wine/drive_c/Program\ Files\ (x86)\/Microsoft\ Visual\ Studio/2017/BuildTools/VC/Tools/MSVC/14.16.27023/bin/Hostx64/x86/bin/link.exe /lib /machine:i386 /def:libhamlib-4.def )
-
+# Path for 2019 version Visual Studio Community
+#( cd ${ZIP_DIR}/lib/msvc/ && wine ~/.wine/drive_c/Program\ Files\ \(x86)\\/Microsoft\ Visual\ Studio\2019\Community\VC\Tools\MSVC\14.25.28610\bin\Hostx64\x86\bin\link.exe /lib /machine:i386 /def:libhamlib-4.def )
 /usr/bin/zip -r ${HL_FILENAME}.zip `basename ${ZIP_DIR}`
